+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- pluck(predicate?) helper to extract a part of the haystack
 - instanceOf(constructor) helper
 
 ## [1.9.0] - 2021-10-23